retest

+++ app.py
@@ -15,7 +15,7 @@ CORS(app)
 class User(db.Model):
     __tablename__ = 'user'
 
-    user_id = db.Column(db.Integer(11), primary_key=True, auto_increment=True)
+    user_id = db.Column(db.Integer, primary_key=True, auto_increment=True)
     user_name = db.Column(db.Varchar(50), nullable=False)
     name = db.Column(db.Varchar(50), nullable=False)
     designation = db.Column(db.Varchar(50), nullable=False)
@@ -78,7 +78,7 @@ class User(db.Model):
 class Course(db.Model):
     __tablename__ = 'course'
 
-    course_id = db.Column(db.Integer(11), primary_key=True, auto_increment=True)
+    course_id = db.Column(db.Integer, primary_key=True, auto_increment=True)
     course_name = db.Column(db.Varchar(50), nullable=False)
     archive_date = db.Column(db.Date)
 
@@ -97,9 +97,9 @@ class Course(db.Model):
 class Prerequisites(db.Model):
     __tablename__ = 'prerequisites'
 
-    prereq_id = db.Column(db.Integer(11), primary_key=True, auto_increment=True)
-    course_id = db.Column(db.Integer(11), db.ForeignKey('course.course_id'), nullable=False)
-    prereq_course_id = db.Column(db.Integer(11), db.ForeignKey('course.course_id'), nullable=False)
+    prereq_id = db.Column(db.Integer, primary_key=True, auto_increment=True)
+    course_id = db.Column(db.Integer, db.ForeignKey('course.course_id'), nullable=False)
+    prereq_course_id = db.Column(db.Integer, db.ForeignKey('course.course_id'), nullable=False)
 
     def to_dict(self):
         columns = self.__mapper__.column_attrs.keys()
@@ -111,11 +111,11 @@ class Prerequisites(db.Model):
 class Class(db.Model):
     __tablename__ = 'class'
 
-    class_id = db.Column(db.Integer(11), primary_key=True, auto_increment=True)
-    course_id = db.Column(db.Integer(11), db.ForeignKey('course.course_id'), nullable=False)
-    trainer_id = db.Column(db.Integer(11), db.ForeignKey('user.user_id'), nullable=False) 
+    class_id = db.Column(db.Integer, primary_key=True, auto_increment=True)
+    course_id = db.Column(db.Integer, db.ForeignKey('course.course_id'), nullable=False)
+    trainer_id = db.Column(db.Integer, db.ForeignKey('user.user_id'), nullable=False) 
     class_name = db.Column(db.Varchar(50), nullable=False)
-    capacity = db.Column(db.Integer(11), nullable=False)
+    capacity = db.Column(db.Integer, nullable=False)
     start_Date = db.Column(db.Date, nullable=False)
     end_Date = db.Column(db.Date, nullable=False)
     start_Time = db.Column(db.Time, nullable=False)
@@ -143,10 +143,10 @@ class Class(db.Model):
 class Chapter(db.Model):
     __tablename__ = 'chapter'
 
-    chapter_id = db.Column(db.Integer(11), primary_key=True, auto_increment=True)
-    class_id = db.Column(db.Integer(11), db.ForeignKey('class.class_id'), nullable=False)
+    chapter_id = db.Column(db.Integer, primary_key=True, auto_increment=True)
+    class_id = db.Column(db.Integer, db.ForeignKey('class.class_id'), nullable=False)
     chapter_name = db.Column(db.Varchar(50), nullable=False)
-    order = db.Column(db.Integer(11), nullable=False)
+    order = db.Column(db.Integer, nullable=False)
     chapter_materials = db.Column(db.Varchar(50), nullable=False)
 
     def to_dict(self):
@@ -159,11 +159,11 @@ class Chapter(db.Model):
 class Quiz(db.Model):
     __tablename__ = 'quiz'
 
-    quiz_id = db.Column(db.Integer(11), primary_key=True, auto_increment=True)
-    chapter_id = db.Column(db.Integer(11), db.ForeignKey('chapter.chapter_id'), nullable=False)
-    duration = db.Column(db.Integer(11), nullable=False)
+    quiz_id = db.Column(db.Integer, primary_key=True, auto_increment=True)
+    chapter_id = db.Column(db.Integer, db.ForeignKey('chapter.chapter_id'), nullable=False)
+    duration = db.Column(db.Integer, nullable=False)
     graded = db.Column(db.Boolean, nullable=False)
-    passing_mark = db.Column(db.Integer(11))
+    passing_mark = db.Column(db.Integer)
 
     def to_dict(self):
         columns = self.__mapper__.column_attrs.keys()
@@ -195,10 +195,10 @@ class Quiz(db.Model):
 class Question(db.Model):
     __tablename__ = 'question'
 
-    question_id = db.Column(db.Integer(11), primary_key=True, auto_increment=True)
-    quiz_id = db.Column(db.Integer(11), db.ForeignKey('quiz.quiz_id'), nullable=False)
+    question_id = db.Column(db.Integer, primary_key=True, auto_increment=True)
+    quiz_id = db.Column(db.Integer, db.ForeignKey('quiz.quiz_id'), nullable=False)
     question = db.Column(db.Varchar(250), nullable=False)
-    marks = db.Column(db.Integer(11), nullable=False)
+    marks = db.Column(db.Integer, nullable=False)
 
     def to_dict(self):
         columns = self.__mapper__.column_attrs.keys()
@@ -222,7 +222,7 @@ class Question(db.Model):
 class Questiontf(db.Model):
     __tablename__ = 'question_tf'
 
-    question_tf_id = db.Column(db.Integer(11), db.ForeignKey('question.question_id'), primary_key=True, nullable=False)
+    question_tf_id = db.Column(db.Integer, db.ForeignKey('question.question_id'), primary_key=True, nullable=False)
     corrected_value = db.Column(db.Boolean, nullable=False)
 
     def to_dict(self):
@@ -239,7 +239,7 @@ class Questiontf(db.Model):
 class Questionmcq(db.Model):
     __tablename__ = 'question_mcq'
 
-    question_mcq_id = db.Column(db.Integer(11), db.ForeignKey('question.question_id'), primary_key=True, nullable=False)
+    question_mcq_id = db.Column(db.Integer, db.ForeignKey('question.question_id'), primary_key=True, nullable=False)
 
     def to_dict(self):
         columns = self.__mapper__.column_attrs.keys()
@@ -256,8 +256,8 @@ class Questionmcq(db.Model):
 class Options(db.Model):
     __tablename__ = 'options'
 
-    options_id = db.Column(db.Integer(11), primary_key=True, nullable=False)
-    question_mcq_id = db.Column(db.Integer(11), db.ForeignKey('question.question_id'), nullable=False)
+    options_id = db.Column(db.Integer, primary_key=True, nullable=False)
+    question_mcq_id = db.Column(db.Integer, db.ForeignKey('question.question_id'), nullable=False)
     value = db.Column(db.Varchar(50), nullable=False)
     corrected_value = db.Column(db.Boolean, nullable=False)
 
@@ -271,11 +271,11 @@ class Options(db.Model):
 class CourseProgression(db.Model):
     __tablename__ = 'course_progression'
 
-    cc_id = db.Column(db.Integer(11), primary_key=True, auto_increment=True)
-    user_id = db.Column(db.Integer(11), db.ForeignKey('user.user_id'), nullable=False)
-    course_id = db.Column(db.Integer(11), db.ForeignKey('course.course_id'), nullable=False)    
-    class_id = db.Column(db.Integer(11), db.ForeignKey('class.class_id'), nullable=False)
-    chapter_id = db.Column(db.Integer(11), db.ForeignKey('chapter.chapter_id'), nullable=False)
+    cc_id = db.Column(db.Integer, primary_key=True, auto_increment=True)
+    user_id = db.Column(db.Integer, db.ForeignKey('user.user_id'), nullable=False)
+    course_id = db.Column(db.Integer, db.ForeignKey('course.course_id'), nullable=False)    
+    class_id = db.Column(db.Integer, db.ForeignKey('class.class_id'), nullable=False)
+    chapter_id = db.Column(db.Integer, db.ForeignKey('chapter.chapter_id'), nullable=False)
     status = db.Column(db.Varchar(50), nullable=False)
     completion_date = db.Column(db.Date)
     score = db.Column(db.Integer)
